+++ ChangeLog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) without the patch version.
 
+## [0.26 - 2023-06-15]
+### Changed
+- Corrected an exception in `VISADevice.list_resources`
+- Set __doc__ for traits, and remove the list of traits from Device __init__ docstrings
+
 ## [0.25 - 2023-06-08]
 ### Added